Fix: change constructor in cont. resource tests
- Continuous resource tests used discrete constructor -> false errors
- Checked for invalid names and units when names and units are coerced
  into strings, so can never be invalid
- Refactored filenames so resource tests can be discovered
- Changes resource to use ClassOfSupply enum, moved to
  spacenet.constants
